requirements.txt
gradio
ultralytics
opencv-python-headless
pillow
torch
torchvision
numpy

